Sets up separate files for save data
