+++ Danki2/Assets/Scripts/Persistence/SaveData.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: f8fbdd8e0d734cbd9783324bceac0f0a
+timeCreated: 1615197246
