+++ Danki2/Assets/Scripts/Persistence/SaveData/CombatSceneSaveData.cs
@@ -1,0 +1,23 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+
+public class CombatSceneSaveData
+{
+    public bool EnemiesCleared { get; set; }
+    public Dictionary<int, ActorType> SpawnerIdToSpawnedActor { get; set; }
+
+    public SerializableCombatSceneSaveData Serialize()
+    {
+        return new SerializableCombatSceneSaveData
+        {
+            EnemiesCleared = EnemiesCleared,
+            SerializableSpawnerSaveDataList = SpawnerIdToSpawnedActor.Keys
+                .Select(spawnerId => new SerializableSpawnerSaveData
+                {
+                    Id = spawnerId,
+                    ActorType = SpawnerIdToSpawnedActor[spawnerId]
+                })
+                .ToList()
+        };
+    }
+}
