+++ Danki2/Assets/Scripts/Persistence/SaveData/CombatSceneSaveData.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 28b162b8bd294b40b03438e74ed8a9d2
+timeCreated: 1615197282
