+++ Danki2/Assets/Scripts/Persistence/SaveData/SceneSaveData.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: dac163fd1eb3468abbfdfdaa32e66b00
+timeCreated: 1615197272
