+++ Danki2/Assets/Scripts/Persistence/SaveData/SceneType.cs
@@ -1,0 +1,6 @@
+﻿public enum SceneType
+{
+    Combat,
+    Shop,
+    Victory
+}
