+++ Danki2/Assets/Scripts/Persistence/SaveData/SceneType.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: f77d3c6d2002428d878738ccde884261
+timeCreated: 1615197294
