+++ Danki2/Assets/Scripts/Persistence/SaveData/Serialization.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 31428a22825548a58f821b72f59815bc
+timeCreated: 1615197305
