+++ Danki2/Assets/Scripts/Persistence/SaveData/Serialization/SerializableCombatSceneSaveData.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 6a04757105d449dea419cbcd0229eb4e
+timeCreated: 1615197376
