+++ Danki2/Assets/Scripts/Persistence/SaveData/Serialization/SerializableSceneTransition.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using UnityEngine;
+
+[Serializable]
+public class SerializableSceneTransition
+{
+    [SerializeField] private int fromId;
+    [SerializeField] private List<int> toIds;
+
+    public int FromId { get => fromId; set => fromId = value; }
+    public List<int> ToIds { get => toIds; set => toIds = value; }
+}
