+++ Danki2/Assets/Scripts/Persistence/SaveData/Serialization/SerializableSceneTransition.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 0f68be95bfec4333bf6b11bb12b16044
+timeCreated: 1615197398
