+++ Danki2/Assets/Scripts/Persistence/SaveData/Serialization/SerializableSpawnerSaveData.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using UnityEngine;
+
+[Serializable]
+public class SerializableSpawnerSaveData
+{
+    [SerializeField] private int id;
+    [SerializeField] private ActorType actorType;
+
+    public int Id { get => id; set => id = value; }
+    public ActorType ActorType { get => actorType; set => actorType = value; }
+}
