+++ Danki2/Assets/Scripts/Persistence/SaveData/Serialization/SerializableSpawnerSaveData.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 50869062cc974c1cae023e1ac5f37bfe
+timeCreated: 1615197387
